Add a profile style guide page
I picked a random profile from people.wsu.edu that has both a
photo and a reasonable amount of personal biography content.

I decided to grab from people.wsu.edu instead of the current theme
output so we can start with a clean slate. Seemed like a good idea
at the time... An HTML comment with all the meta data a profile
can include is provided so we can sort out how it should all be
handled. I also included the theme header in case we decide to use
it.