Allow users with the Profile Owner role to edit their profiles
I mistakenly thought this was already working, but my local testing
was flawed because my test user was the author of my test profile.
Such cases are the exception on the production site.

Also, I couldn't for the life of me get this to work the other way
around (that is, granting users the `edit_others_profiles` capability
for the profile with a NID that matches their login). That would have
been the preferable route...