Update alternative field values with new AD data if applicable
Because I made weird decisions, the AD data is also stored as the
value for the alternative fields until the latter are changed by
the user. This accommodates for the possibility that the AD value
is still in place for the alternative fields and updates them with
new AD data if any is available.